# ballgamemanage/加了swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "swagger学习文档，这条非必须",
        "title": "球赛管理系统",
        "contact": {},
        "version": "最后版未完成"
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/games": {
            "get": {
                "description": "查询all games信息并返回到客户端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "查询game",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Game"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "接收game信息然后创建并返回到客户端反馈",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "game"
                ],
                "summary": "创建球赛",
                "responses": {
                    "200": {
                        "description": "new game创建成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/players": {
            "post": {
                "description": "接收players信息然后登记创建并返回到客户端反馈",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "登记运动员",
                "responses": {
                    "200": {
                        "description": "new player创建成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/players/{playerid}": {
            "get": {
                "description": "查询指定id的player信息并返回到客户端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "查询player",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Player"
                        }
                    }
                }
            }
        },
        "/teames": {
            "post": {
                "description": "接收teams信息然后登记创建并返回到客户端反馈",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "登记队伍",
                "responses": {
                    "200": {
                        "description": "new team创建成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/teams/{playerid}": {
            "put": {
                "description": "更新修改运动员的team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player"
                ],
                "summary": "更新运动员的team信息",
                "responses": {
                    "200": {
                        "description": "player的newteamid修改成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/teams/{teamid}": {
            "get": {
                "description": "查询指定id的team信息并返回到客户端",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "查询team",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    }
                }
            }
        },
        "/teams/{teamid}/{newinfo}": {
            "put": {
                "description": "更新team的info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "更新team的info",
                "responses": {
                    "200": {
                        "description": "修改teaminfo成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "获取newinfo错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/teams/{teamid}/{newlogo}": {
            "put": {
                "description": "更新team的logo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "更新team的logo",
                "responses": {
                    "200": {
                        "description": "修改teamlogo成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "获取newlogo错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/teams/{teamid}/{newname}": {
            "put": {
                "description": "更新team的name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "更新team的name",
                "responses": {
                    "200": {
                        "description": "player的newteamid修改成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/teams/{usename/{newrole}": {
            "put": {
                "description": "更改用户权限role",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "更改用户权限",
                "responses": {}
            }
        },
        "/users": {
            "post": {
                "description": "接收user信息然后创建并返回到客户端反馈",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "创建用户",
                "responses": {
                    "200": {
                        "description": "new user创建成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/appoint/{username}": {
            "get": {
                "description": "查询用户预约的比赛详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "查询用户预约",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserAppoint"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "接收user预约信息然后创建并返回到客户端反馈",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appoint"
                ],
                "summary": "创建用户预约",
                "responses": {
                    "200": {
                        "description": "new appoint创建成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error\": err.Error()}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Game": {
            "type": "object",
            "properties": {
                "appointment": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                }
            }
        },
        "model.Player": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "model.Team": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.UserAppoint": {
            "type": "object",
            "properties": {
                "appoint_game_name": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}